Remove most of "unsupported browser" logic from the app; No mobile browsers should be allowed; Implement the mobile browsers support only for the oAuth case;
