+++ src/script/auth/page/ConversationJoin.tsx
@@ -40,8 +40,8 @@ import {Page} from './Page';
 import {Config} from '../../Config';
 import {conversationJoinStrings} from '../../strings';
 import {AppAlreadyOpen} from '../component/AppAlreadyOpen';
+import {HasToUseDesktop} from '../component/HasToUseDesktop';
 import {JoinGuestLinkPasswordModal} from '../component/JoinGuestLinkPasswordModal';
-import {UnsupportedBrowser} from '../component/UnsupportedBrowser';
 import {WirelessContainer} from '../component/WirelessContainer';
 import {EXTERNAL_ROUTE} from '../externalRoute';
 import {actionRoot as ROOT_ACTIONS} from '../module/action';
@@ -251,7 +251,7 @@ const ConversationJoinComponent = ({
   }
 
   return (
-    <UnsupportedBrowser isTemporaryGuest>
+    <HasToUseDesktop isTemporaryGuest>
       {isJoinGuestLinkPasswordModalOpen && (
         <JoinGuestLinkPasswordModal
           onClose={() => {
@@ -308,7 +308,7 @@ const ConversationJoinComponent = ({
           </Column>
         </Columns>
       </WirelessContainer>
-    </UnsupportedBrowser>
+    </HasToUseDesktop>
   );
 };
