+++ src/script/auth/page/Login.test.tsx
@@ -43,17 +43,7 @@ describe('Login', () => {
 
     spyOn(actionRoot.authAction, 'doLogin').and.returnValue(() => Promise.resolve());
 
-    const {getByTestId} = mountComponent(
-      <Login />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
     const emailInput = getByTestId('enter-email');
     const passwordInput = getByTestId('enter-password');
@@ -83,17 +73,7 @@ describe('Login', () => {
       Promise.reject(new BackendError('Too many clients', BackendErrorLabel.TOO_MANY_CLIENTS, StatusCodes.NOT_FOUND)),
     );
 
-    const {getByTestId} = mountComponent(
-      <Login />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
     const emailInput = getByTestId('enter-email');
     const passwordInput = getByTestId('enter-password');
@@ -121,17 +101,7 @@ describe('Login', () => {
 
     spyOn(actionRoot.authAction, 'doLogin').and.returnValue(() => Promise.resolve());
 
-    const {getByTestId} = mountComponent(
-      <Login />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
     const emailInput = getByTestId('enter-email');
     const passwordInput = getByTestId('enter-password');
@@ -152,17 +122,7 @@ describe('Login', () => {
   });
 
   it('has disabled submit button as long as one input is empty', () => {
-    const {getByTestId} = mountComponent(
-      <Login />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
     const emailInput = getByTestId('enter-email');
     const passwordInput = getByTestId('enter-password');
@@ -186,17 +146,7 @@ describe('Login', () => {
           ENABLE_SSO: false,
         },
       });
-      const {queryByTestId} = mountComponent(
-        <Login />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {queryByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
       const backButton = queryByTestId('go-index');
       expect(backButton).toBeNull();
@@ -210,17 +160,7 @@ describe('Login', () => {
           ENABLE_ACCOUNT_REGISTRATION: true,
         },
       });
-      const {getByTestId} = mountComponent(
-        <Login />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {getByTestId} = mountComponent(<Login />, mockStoreFactory()(initialRootState));
 
       const backButton = getByTestId('go-index');
       expect(backButton).not.toBeNull();
