+++ src/script/auth/page/SetAccountType.test.tsx
@@ -51,17 +51,7 @@ describe('when visiting the set account type page', () => {
     });
 
     it('redirects to the index page', () => {
-      const {getByText} = mountComponent(
-        <SetAccountType />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {getByText} = mountComponent(<SetAccountType />, mockStoreFactory()(initialRootState));
       const redirect = getByText('Navigate to /');
 
       expect(redirect).not.toBeNull();
@@ -81,51 +71,21 @@ describe('when visiting the set account type page', () => {
     });
 
     it('shows the Wire logo', () => {
-      const {getByTestId} = mountComponent(
-        <SetAccountType />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {getByTestId} = mountComponent(<SetAccountType />, mockStoreFactory()(initialRootState));
       const logo = getByTestId(logoId);
 
       expect(logo).not.toBeNull();
     });
 
     it('shows an option to create a private account', () => {
-      const {getByTestId} = mountComponent(
-        <SetAccountType />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {getByTestId} = mountComponent(<SetAccountType />, mockStoreFactory()(initialRootState));
       const personalAccountButton = getByTestId(personalAccountButtonId);
 
       expect(personalAccountButton).not.toBeNull();
     });
 
     it('shows an option to create a team', () => {
-      const {getByTestId} = mountComponent(
-        <SetAccountType />,
-        mockStoreFactory()({
-          ...initialRootState,
-          runtimeState: {
-            hasCookieSupport: true,
-            hasIndexedDbSupport: true,
-            isSupportedBrowser: true,
-          },
-        }),
-      );
+      const {getByTestId} = mountComponent(<SetAccountType />, mockStoreFactory()(initialRootState));
 
       const teamAccountButton = getByTestId(teamAccountButtonId);
       expect(teamAccountButton).not.toBeNull();
