+++ src/script/auth/page/SetHandle.test.tsx
@@ -34,17 +34,7 @@ describe('SetHandle', () => {
   it('has disabled submit button as long as there is no input', async () => {
     spyOn(actionRoot.selfAction, 'doGetConsents').and.returnValue(() => Promise.resolve());
     spyOn(actionRoot.userAction, 'checkHandles').and.returnValue(() => Promise.resolve());
-    const {getByTestId} = mountComponent(
-      <SetHandle />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<SetHandle />, mockStoreFactory()(initialRootState));
 
     await waitFor(() => getByTestId(handleInputId));
     const handleInput = getByTestId(handleInputId);
@@ -63,17 +53,7 @@ describe('SetHandle', () => {
 
     const handle = 'handle';
 
-    const {getByTestId} = mountComponent(
-      <SetHandle />,
-      mockStoreFactory()({
-        ...initialRootState,
-        runtimeState: {
-          hasCookieSupport: true,
-          hasIndexedDbSupport: true,
-          isSupportedBrowser: true,
-        },
-      }),
-    );
+    const {getByTestId} = mountComponent(<SetHandle />, mockStoreFactory()(initialRootState));
 
     await waitFor(() => getByTestId(handleInputId));
     const handleInput = getByTestId(handleInputId);
